Added code to attach to the default interface
When attaching to the real interface we need to check we are setting the
right interface in eucalyptus.conf for the CC, and that a set of
PUBLIC_IPS is correctly set for the instances to use.

Right now there is very little checking, but the main interface (the one
with the default route) is identified and used. The IP of the interface is
the one used to register the components.